Golden Code, before GUI. Almost all functionality included

+++ Atmel Studio/7.0/MEGA_UART_INTERRUPT_EXAMPLE1/MEGA_GPIO_EXAMPLE1/src/adc_func.h
@@ -35,6 +35,6 @@ float Amp_ADC(int Amp_num);
 float Temp_ADC();
 
 // Get state of charge
-int SoC_ADC(float volt, float amp);
+float SoC_ADC(float volt, float amp);
 
 #endif /* ADC_FUNC_H_ */
